Update hardhat deploy scripts
