+++ apps/smart-contracts/package.json
@@ -15,6 +15,7 @@
   "type": "commonjs",
   "devDependencies": {
     "@nomicfoundation/hardhat-toolbox": "^5.0.0",
+    "dotenv": "^16.4.7",
     "hardhat": "^2.22.19",
     "typescript": "^5.8.3"
   },
